requirements.txt
requests
beautifulsoup4
anthropic
python-telegram-bot
